refactor: :art: folder structure
