+++ CHANGELOG.md
@@ -3,6 +3,8 @@
 ## v0.1.0 (No lanzada)
 
 ### Nuevo
+-		Organización de la estructura de carpetas.
+-		Uso de la libreria `color_eyre`.
 -   Añadido el comando `!ping`.
--   Uso de la librería `Serenity`.
+-   Uso de la librería `serenity`.
 -   Fase inicial del proyecto.
