+++ src/commands/mod.rs
@@ -1,0 +1,8 @@
+pub mod ping;
+
+use ping::*;
+use serenity::framework::standard::macros::group;
+
+#[group]
+#[commands(ping)]
+struct General;
